# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"","version":"0.0.1"},"host":"localhost:80","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/tasks/{taskId}/complete":{"patch":{"tags":["mktextr"],"summary":"completeTask mktextr","operationId":"mktextr#completeTask","consumes":["multipart/form-data"],"parameters":[{"name":"taskId","in":"path","description":"ID of the task","required":true,"type":"string"},{"name":"CompleteTaskRequestBody","in":"formData","required":true,"schema":{"$ref":"#/definitions/MktextrCompleteTaskRequestBody","required":["file","filename"]}}],"responses":{"200":{"description":"OK response."}},"schemes":["http"]}},"/textures":{"get":{"tags":["mktextr"],"summary":"getTextureByCoordinates mktextr","operationId":"mktextr#getTextureByCoordinates","parameters":[{"name":"worldId","in":"query","description":"WorldId","required":true,"type":"string"},{"name":"x","in":"query","description":"Texture X","required":true,"type":"integer"},{"name":"y","in":"query","description":"Texture Y","required":true,"type":"integer"}],"responses":{"202":{"description":"Accepted response.","schema":{"$ref":"#/definitions/MktextrGetTextureByCoordinatesAcceptedResponseBody"},"headers":{"X-mktextr-task-id":{"type":"string"}}},"308":{"description":"Permanent Redirect response.","schema":{"$ref":"#/definitions/MktextrGetTextureByCoordinatesPermanentRedirectResponseBody"},"headers":{"Location":{"type":"string"}}},"500":{"description":"Internal Server Error response.","schema":{"$ref":"#/definitions/MktextrGetTextureByCoordinatesInternalServerErrorResponseBody"}}},"schemes":["http"]}},"/textures/{id}":{"get":{"tags":["mktextr"],"summary":"getTextureById mktextr","operationId":"mktextr#getTextureById","parameters":[{"name":"id","in":"path","description":"Texture ID","required":true,"type":"string"}],"responses":{"204":{"description":"No Content response."}},"schemes":["http"]}}},"definitions":{"MktextrCompleteTaskRequestBody":{"title":"MktextrCompleteTaskRequestBody","type":"object","properties":{"file":{"type":"string","description":"The file to upload","example":"VmVsaXQgcGxhY2VhdC4=","format":"byte"},"filename":{"type":"string","description":"Name of the file","example":"Voluptate dolorem quia."}},"example":{"file":"U2FlcGUgYXRxdWUgdXQu","filename":"Quos est ad ut voluptatibus."},"required":["file","filename"]},"MktextrGetTextureByCoordinatesAcceptedResponseBody":{"title":"MktextrGetTextureByCoordinatesAcceptedResponseBody","type":"object","properties":{"Location":{"type":"string","example":"Odio fugit nemo dolorem voluptas consequatur impedit."}},"example":{"Location":"Adipisci consectetur voluptate quia sunt saepe."}},"MktextrGetTextureByCoordinatesInternalServerErrorResponseBody":{"title":"MktextrGetTextureByCoordinatesInternalServerErrorResponseBody","type":"object","properties":{"Location":{"type":"string","example":"Et non cupiditate ut et quasi sit."},"X-mktextr-task-id":{"type":"string","example":"Repellendus perferendis."}},"example":{"Location":"Repudiandae animi.","X-mktextr-task-id":"Ut rerum velit."}},"MktextrGetTextureByCoordinatesPermanentRedirectResponseBody":{"title":"MktextrGetTextureByCoordinatesPermanentRedirectResponseBody","type":"object","properties":{"X-mktextr-task-id":{"type":"string","example":"Eveniet possimus possimus aut quia."}},"example":{"X-mktextr-task-id":"Recusandae reprehenderit et repellat qui et dicta."}}}}